crawler/requirements.txt
aiohttp
beautifulsoup4
lxml
python-multipart
trafilatura
readability-lxml
validators
aiofiles
asyncio
requests
brotlipy